Revert "Auto merge of #130040 - onur-ozkan:llvm-tools-with-ci-rustc, r=Kobzol"
This reverts commit adaff5368b0c7b328a0320a218751d65ab1bba97, reversing
changes made to 2e8db5e9e39c2bf7729113b3041ef4011d90ac5a.
